Move pointer-based implementation to submodule
